+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Release history
 ---------------
 
+0.2.0 (2021-07-30)
+++++++++++++++++++
+
+- AWS credentials usage with rasterio (#19)
+- Rioxarray usage to read data (#18)
+- New StacAssets driver (#17)
+- `get_data()` now accepts the same types of geometries as `eodag.search()` (#16)
+- New notebook and readme example (#20)
+
 0.1.2 (2021-06-18)
 ++++++++++++++++++